Fix problem that would arise upon having a messagebox open and at the same time increasing the number of messages shown in the history.
Revamp the Settings interface at the same time. It now features one signal,
"settings_changed", that passes along the key that have been changed.
Callbacks listening to this signal should then use the key to determine
whether this is the key that they're interested in and check for the new
value if that's the case.